refactor(ui): move shadcn primitives into @ybdownload/ui
Relocate all desktop components/ui primitives to packages/ui with Radix
and lucide deps; desktop keeps thin re-exports and app-specific Toaster.
