+++ apps/desktop/frontend/src/components/ui/button.tsx
@@ -1,5 +1,1 @@
-export {
-  Button,
-  buttonVariants,
-  type ButtonProps,
-} from "@ybdownload/ui/button";
+export * from "@ybdownload/ui/button";
