+++ apps/desktop/frontend/src/components/ui/progress.tsx
@@ -1,26 +1,1 @@
-import * as React from "react";
-import * as ProgressPrimitive from "@radix-ui/react-progress";
-
-import { cn } from "@/lib/utils";
-
-const Progress = React.forwardRef<
-  React.ElementRef<typeof ProgressPrimitive.Root>,
-  React.ComponentPropsWithoutRef<typeof ProgressPrimitive.Root>
->(({ className, value, ...props }, ref) => (
-  <ProgressPrimitive.Root
-    ref={ref}
-    className={cn(
-      "bg-secondary relative h-2 w-full overflow-hidden rounded-full",
-      className
-    )}
-    {...props}
-  >
-    <ProgressPrimitive.Indicator
-      className="bg-primary h-full w-full flex-1 transition-all duration-300 ease-out"
-      style={{ transform: `translateX(-${100 - (value || 0)}%)` }}
-    />
-  </ProgressPrimitive.Root>
-));
-Progress.displayName = ProgressPrimitive.Root.displayName;
-
-export { Progress };
+export * from "@ybdownload/ui/progress";
