+++ apps/desktop/frontend/src/components/ui/toaster.tsx
@@ -1,4 +1,3 @@
-import { useToast } from "@/hooks/use-toast";
 import {
   Toast,
   ToastClose,
@@ -6,7 +5,8 @@ import {
   ToastProvider,
   ToastTitle,
   ToastViewport,
-} from "@/components/ui/toast";
+} from "@ybdownload/ui/toast";
+import { useToast } from "@/hooks/use-toast";
 
 export function Toaster() {
   const { toasts } = useToast();
