+++ packages/ui/package.json
@@ -5,8 +5,23 @@
   "type": "module",
   "exports": {
     ".": "./src/index.ts",
-    "./button": "./src/components/button.tsx",
     "./badge": "./src/components/badge.tsx",
+    "./button": "./src/components/button.tsx",
+    "./card": "./src/components/card.tsx",
+    "./collapsible": "./src/components/collapsible.tsx",
+    "./input": "./src/components/input.tsx",
+    "./label": "./src/components/label.tsx",
+    "./marquee-text": "./src/components/marquee-text.tsx",
+    "./progress": "./src/components/progress.tsx",
+    "./radio-group": "./src/components/radio-group.tsx",
+    "./scroll-area": "./src/components/scroll-area.tsx",
+    "./select": "./src/components/select.tsx",
+    "./separator": "./src/components/separator.tsx",
+    "./switch": "./src/components/switch.tsx",
+    "./tabs": "./src/components/tabs.tsx",
+    "./textarea": "./src/components/textarea.tsx",
+    "./toast": "./src/components/toast.tsx",
+    "./tooltip": "./src/components/tooltip.tsx",
     "./utils": "./src/lib/utils.ts"
   },
   "scripts": {
@@ -14,9 +29,21 @@
     "lint": "tsc --noEmit"
   },
   "dependencies": {
+    "@radix-ui/react-collapsible": "^1.1.13",
+    "@radix-ui/react-label": "^2.1.9",
+    "@radix-ui/react-progress": "^1.1.9",
+    "@radix-ui/react-radio-group": "^1.4.0",
+    "@radix-ui/react-scroll-area": "^1.2.11",
+    "@radix-ui/react-select": "^2.3.0",
+    "@radix-ui/react-separator": "^1.1.9",
     "@radix-ui/react-slot": "^1.1.1",
+    "@radix-ui/react-switch": "^1.3.0",
+    "@radix-ui/react-tabs": "^1.1.14",
+    "@radix-ui/react-toast": "^1.2.16",
+    "@radix-ui/react-tooltip": "^1.2.9",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "lucide-react": "^0.468.0",
     "tailwind-merge": "^3.6.0"
   },
   "devDependencies": {
